Changed calculations to be done on a float value map and am now using esp-dsp

+++ main/computation/include/computation.h
@@ -5,7 +5,7 @@
 esp_err_t init_computation();
 
 /*this stores the frame of the processed microscopic image. It is greyscale image (1 byte per pixel)*/
-extern uint8_t* computation_frame;
+extern float* computation_frame;
 
 /*C_scale for QDF calculations*/
 extern float c_scale;
@@ -19,34 +19,34 @@ void begin_new_computation(void* handler_arg, esp_event_base_t base, int32_t id,
  * @param data greyscale image(s) to use in calculations
  * @param buf buffer to push processed image to.
 */
-void run_bf(uint8_t** data, uint8_t** buf);
+void run_bf(uint8_t** data, float** buf);
 /*
  * @param data greyscale image(s) to use in calculations
  * @param buf buffer to push processed image to.
 */
-void run_df(uint8_t** data, uint8_t** buf);
+void run_df(uint8_t** data, float** buf);
 /*
  * @param data greyscale image(s) to use in calculations
  * @param buf buffer to push processed image to.
 */
-void run_qdf(uint8_t** data, uint8_t** buf, float c_scale);
+void run_qdf(uint8_t** data, float** buf, float c_scale);
 /*
  * @param data greyscale image(s) to use in calculations
  * @param buf buffer to push processed image to.
 */
-void run_dpc_lr(uint8_t** data, uint8_t** buf);
+void run_dpc_lr(uint8_t** data, float** buf);
 /*
  * @param data greyscale image(s) to use in calculations
  * @param buf buffer to push processed image to.
 */
-void run_dpc_rl(uint8_t** data, uint8_t** buf);
+void run_dpc_rl(uint8_t** data, float** buf);
 /*
  * @param data greyscale image(s) to use in calculations
  * @param buf buffer to push processed image to.
 */
-void run_dpc_tb(uint8_t** data, uint8_t** buf);
+void run_dpc_tb(uint8_t** data, float** buf);
 /*
  * @param data greyscale image(s) to use in calculations
  * @param buf buffer to push processed image to.
 */
-void run_dpc_bt(uint8_t** data, uint8_t** buf);
+void run_dpc_bt(uint8_t** data, float** buf);
